+++ src/hw/spi/spi_loc.h
@@ -57,6 +57,6 @@
 #define SPI_COMMAND_WRITE_ENABLE    0x06
 #define SPI_COMMAND_ERASE_SECTOR    0x20
 
-SINT16 hw_spi_xfer (UINT16 nbytes, UINT8 *dataOut, UINT8 *dataIn, spiConfig_t *cfg, BOOL terminate);
+SINT16 hw_spi_xfer (UINT32 nbytes, UINT8 *dataOut, UINT8 *dataIn, spiConfig_t *cfg, BOOL terminate);
 
 #endif /* _SPI_LOC_H */
